# api/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_noah-isme_sma-adp-api_internal_models.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
    type: object
...
  github_com_noah-isme_sma-adp-api_internal_service.BulkCreateSchedulesRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_service.CreateScheduleRequest'
        minItems: 1
        type: array
      partial_on_error:
        type: boolean
    required:
    - items
    type: object
...
  github_com_noah-isme_sma-adp-api_internal_service.CreateScheduleRequest:
    properties:
      class_id:
        type: string
      day_of_week:
        type: string
      room:
        type: string
      subject_id:
        type: string
      teacher_id:
        type: string
      term_id:
        type: string
      time_slot:
        type: string
    required:
    - class_id
    - day_of_week
    - room
    - subject_id
    - teacher_id
    - term_id
    - time_slot
    type: object
...
  github_com_noah-isme_sma-adp-api_internal_service.UpdateScheduleRequest:
    properties:
      class_id:
        type: string
      day_of_week:
        type: string
      room:
        type: string
      subject_id:
        type: string
      teacher_id:
        type: string
      term_id:
        type: string
      time_slot:
        type: string
    required:
    - class_id
    - day_of_week
    - room
    - subject_id
    - teacher_id
    - term_id
    - time_slot
    type: object
...
  github_com_noah-isme_sma-adp-api_pkg_errors.Error:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  github_com_noah-isme_sma-adp-api_pkg_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_errors.Error'
      meta:
        additionalProperties: true
        type: object
      pagination:
        $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_models.Pagination'
    type: object
info:
  contact: {}
  description: Bootstrap server for Golang migration (Phase 0)
  title: SMA ADP API
  version: 0.1.0
paths:
...
  /classes/{id}/schedules:
    get:
      parameters:
      - description: Class ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List schedules by class
      tags:
      - Schedules
...
  /schedules:
    get:
      parameters:
      - description: Filter by term
        in: query
        name: termId
        type: string
      - description: Filter by class
        in: query
        name: classId
        type: string
      - description: Filter by teacher
        in: query
        name: teacherId
        type: string
      - description: Filter by day
        in: query
        name: dayOfWeek
        type: string
      - description: Filter by time slot
        in: query
        name: timeSlot
        type: string
      - description: Filter by room
        in: query
        name: room
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List schedules
      tags:
      - Schedules
    post:
      consumes:
      - application/json
      parameters:
      - description: Schedule payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_service.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Create schedule
      tags:
      - Schedules
  /schedules/{id}:
    delete:
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete schedule
      tags:
      - Schedules
    put:
      consumes:
      - application/json
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      - description: Schedule payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_service.UpdateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Update schedule
      tags:
      - Schedules
  /schedules/bulk:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bulk payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/github_com_noah-isme_sma-adp-api_internal_service.BulkCreateSchedulesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Bulk create schedules
      tags:
      - Schedules
  /schedules/export/pdf:
    get:
      parameters:
      - description: Class ID
        in: query
        name: class_id
        type: string
      - description: Class ID (camelCase)
        in: query
        name: classId
        type: string
      - description: Term ID
        in: query
        name: term_id
        type: string
      - description: Term ID (camelCase)
        in: query
        name: termId
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: Export schedule grid as PDF
      tags:
      - Schedules
...
  /teachers/{id}/schedules:
    get:
      parameters:
      - description: Teacher ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_noah-isme_sma-adp-api_pkg_response.Envelope'
      security:
      - BearerAuth: []
      summary: List schedules by teacher
      tags:
      - Schedules
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"